update query get product

+++ src/data/products_components/product.rs
@@ -5,7 +5,7 @@ pub struct Product {
     pub name: String,
     pub description: Option<String>,
     pub primary_image_id: Option<i32>,
-    pub price: i32,
+    pub price: f32,
     pub stock_quantity: Option<i32>,
     pub size_id: Option<i32>,
     pub category_id: Option<i32>,
